Feat/after

See merge request s11-webmobile4-sub2/S11P12D211!70

+++ android/Signal/app/src/main/java/com/ongo/signal/data/model/chat/ChatHomeCreate.kt
@@ -3,6 +3,6 @@ package com.ongo.signal.data.model.chat
 import com.google.gson.annotations.SerializedName
 
 data class ChatHomeCreate(
-    @SerializedName("from_id") val fromId : Int,
-    @SerializedName("to_id")val toId : Int,
+    @SerializedName("from_id") val fromId : Long,
+    @SerializedName("to_id")val toId : Long,
 )
